Implement Logger as AcraWrapper
